Fix datatype bug, add new shooter motors to code

+++ scripts/behaviors/behaviorlib.py
@@ -67,13 +67,13 @@ def computeShootOffCommand():
 def computeFeedOnCommand():
     twist_cmd = None # Will retain previous value
     shooter_cmd = None
-    feed_cmd = Int32()
+    feed_cmd = Int16()
     feed_cmd.data = 50
     return twist_cmd, shooter_cmd, feed_cmd
 def computeFeedOffCommand():
     twist_cmd = None # Will retain previous value
     shooter_cmd = None
-    feed_cmd = Int32()
+    feed_cmd = Int16()
     feed_cmd.data = 0
     return twist_cmd, shooter_cmd, feed_cmd
